ci: add missing files
