+++ .github/workflows/tests.yml
@@ -14,7 +14,7 @@ jobs:
       - name: Setups Python
         uses: actions/setup-python@v4.2.0
         with:
-          python-version: "3.10"
+          python-version: "3.9"
       - name: Setup Poetry
         uses: abatilo/actions-poetry@v2.2.0
       - name: Run Pre-Commit
